Started refactoring linear algebra.
